+++ qualification-svc/src/main/java/com/example/qualificationsvc/endpoint/rest/api/QualificationController.java
@@ -1,5 +1,6 @@
 package com.example.qualificationsvc.endpoint.rest.api;
 
+import com.example.qualificationsvc.domain.service.AuthorizationService;
 import com.example.qualificationsvc.domain.service.ProjectInfoService;
 import com.example.qualificationsvc.domain.service.QualificationService;
 import com.example.qualificationsvc.domain.service.SkillService;
@@ -10,12 +11,15 @@ import com.example.search.endpoint.rest.api.DefaultApi;
 import com.example.search.endpoint.rest.dto.ProjectRequestDto;
 import com.example.search.endpoint.rest.dto.QualificationProfileDto;
 import com.example.search.endpoint.rest.dto.SkillRequestDto;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 
 @Controller
 @RequiredArgsConstructor
+@SecurityRequirement(name = "bearerAuth")
 public class QualificationController implements DefaultApi {
 
     private final ProjectInfoService projectInfoService;
@@ -24,41 +28,75 @@ public class QualificationController implements DefaultApi {
     private final QualificationDtoMapper qualificationDtoMapper;
     private final SkillDtoMapper skillDtoMapper;
     private final ProjectInfoDtoMapper projectInfoDtoMapper;
+    private final AuthorizationService authorizationService;
+
 
     @Override
     public ResponseEntity<QualificationProfileDto> qualificationPersonIdGet(String personId) {
-        return ResponseEntity.ok(
-              qualificationDtoMapper.toDto(qualificationService.findById(personId)));
-
+        if (authorizationService.isAdmin() ||
+              authorizationService.isHR() ||
+              authorizationService.isManager() ||
+              authorizationService.isEligibleUser(personId)) {
+            return ResponseEntity.ok(
+                  qualificationDtoMapper.toDto(qualificationService.findById(personId)));
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @Override
     public ResponseEntity<Void> qualificationProjectPost(ProjectRequestDto projectRequestDto) {
-        projectInfoService.addProject(projectInfoDtoMapper.toDomain(projectRequestDto));
-        return ResponseEntity.noContent().build();
+        if (authorizationService.isAdmin() || authorizationService.isManager()
+              || authorizationService.isEligibleUser(projectRequestDto.getPersonId())) {
+            projectInfoService.addProject(projectInfoDtoMapper.toDomain(projectRequestDto));
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
+    //TODO resolve delete by user
     @Override
     public ResponseEntity<Void> qualificationProjectProjectIdDelete(String projectId) {
-        projectInfoService.deleteProject(projectId);
-        return ResponseEntity.noContent().build();
+        if (authorizationService.isAdmin() || authorizationService.isManager()) {
+            projectInfoService.deleteProject(projectId);
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
     }
 
     @Override
     public ResponseEntity<Void> qualificationProjectPut(ProjectRequestDto projectRequestDto) {
-        projectInfoService.updateProject(projectInfoDtoMapper.toDomain(projectRequestDto));
-        return ResponseEntity.noContent().build();
+        if (authorizationService.isAdmin() || authorizationService.isManager()
+              || authorizationService.isEligibleUser(projectRequestDto.getPersonId())) {
+            projectInfoService.updateProject(projectInfoDtoMapper.toDomain(projectRequestDto));
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
     }
 
     @Override
     public ResponseEntity<Void> qualificationSkillPost(SkillRequestDto skillRequestDto) {
-        skillService.addSkill(skillDtoMapper.toDomain(skillRequestDto));
-        return ResponseEntity.noContent().build();
+        if (authorizationService.isAdmin() || authorizationService.isManager()
+              || authorizationService.isEligibleUser(skillRequestDto.getPersonId())) {
+            skillService.addSkill(skillDtoMapper.toDomain(skillRequestDto));
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
     }
 
+    //TODO resolve delete by user
     @Override
     public ResponseEntity<Void> qualificationSkillSkillIdDelete(String skillId) {
-        skillService.deleteSkill(skillId);
-        return ResponseEntity.noContent().build();
+        if (authorizationService.isAdmin() || authorizationService.isManager()) {
+            skillService.deleteSkill(skillId);
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
     }
 }
